Updated types and actions
make Containers allow mixed items, instead of having a generic type that makes them all take the same type of item
extract TokenContents to its own type, for consistency
changed TokenType to TokenDefinition, for consistency
removed the ability to put tokens on cards (This seems complicated, and I think we can mostly do this by putting tokens in the same pile as cards)
removed isTemporary flag for items (since we are now just going to delete items permanently via a separate action,
	rather than basing this on whether an individual item is marked as temporary or permanent
refactored containers to a new page
combined cardPool and tokenTypes into itemPool
Updated actions to be generic over all types of item/container when reasonable.
Added actions to interact with the counts of counters
Changed actions for interacting with definitions to fit with the new system (copying cards and adding cards to the card pool are separated)
The new actions are now all categorized under ItemAction. Feel free to add your own subsets of these if they end up being useful.